src/main/java/Staff/Employee.java
package Staff;

public class Employee {
    private String name;
    private String niNumber;
    private double salary;

    public Employee(String name, String niNumber, double salary){
        this.name = name;
        this.niNumber = niNumber;
        this.salary = salary;

    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        if(!(name == null)) {
            this.name = name;
        }
    }

    public String getNiNumber() {
        return niNumber;
    }

    public void setNiNumber(String niNumber) {
        this.niNumber = niNumber;
    }

    public double getSalary() {
        return salary;
    }

    public void setSalary(double salary) {
        this.salary = salary;
    }

    public void raiseSalary(double increase){
        if(increase > 0) {
            this.salary += increase;
        }

    }

    public double payBonus(){
        return this.salary * 0.01;
    }
}
